+++ fmriprep/workflows/anatomical.py
@@ -43,7 +43,7 @@ def init_anat_preproc_wf(skull_strip_ants, output_spaces, template, debug, frees
     outputnode = pe.Node(niu.IdentityInterface(
         fields=['t1_preproc', 't1_brain', 't1_mask', 't1_seg', 't1_tpms',
                 't1_2_mni', 't1_2_mni_forward_transform', 't1_2_mni_reverse_transform',
-                'mni_mask', 'mni_tpms',
+                'mni_mask', 'mni_seg', 'mni_tpms',
                 'subjects_dir', 'subject_id', 'fs_2_t1_transform', 'surfaces']),
         name='outputnode')
 
@@ -96,6 +96,12 @@ def init_anat_preproc_wf(skull_strip_ants, output_spaces, template, debug, frees
         name='mni_mask'
     )
 
+    mni_seg = pe.Node(
+        ants.ApplyTransforms(dimension=3, default_value=0, float=True,
+                             interpolation='NearestNeighbor'),
+        name='mni_seg'
+    )
+
     mni_tpms = pe.MapNode(
         ants.ApplyTransforms(dimension=3, default_value=0, float=True,
                              interpolation='Linear'),
@@ -121,6 +127,7 @@ def init_anat_preproc_wf(skull_strip_ants, output_spaces, template, debug, frees
 
         t1_2_mni.inputs.template = template_str
         mni_mask.inputs.reference_image = ref_img
+        mni_seg.inputs.reference_image = ref_img
         mni_tpms.inputs.reference_image = ref_img
 
         workflow.connect([
@@ -128,6 +135,8 @@ def init_anat_preproc_wf(skull_strip_ants, output_spaces, template, debug, frees
             (skullstrip_wf, t1_2_mni, [('outputnode.out_mask', 'moving_mask')]),
             (skullstrip_wf, mni_mask, [('outputnode.out_mask', 'input_image')]),
             (t1_2_mni, mni_mask, [('composite_transform', 'transforms')]),
+            (t1_seg, mni_seg, [('tissue_class_map', 'input_image')]),
+            (t1_2_mni, mni_seg, [('composite_transform', 'transforms')]),
             (t1_seg, mni_tpms, [('probability_maps', 'input_image')]),
             (t1_2_mni, mni_tpms, [('composite_transform', 'transforms')]),
             (t1_2_mni, outputnode, [
@@ -135,6 +144,7 @@ def init_anat_preproc_wf(skull_strip_ants, output_spaces, template, debug, frees
                 ('composite_transform', 't1_2_mni_forward_transform'),
                 ('inverse_composite_transform', 't1_2_mni_reverse_transform')]),
             (mni_mask, outputnode, [('output_image', 'mni_mask')]),
+            (mni_seg, outputnode, [('output_image', 'mni_seg')]),
             (mni_tpms, outputnode, [('output_image', 'mni_tpms')]),
         ])
 
@@ -193,9 +203,11 @@ def init_anat_preproc_wf(skull_strip_ants, output_spaces, template, debug, frees
             ('t1_preproc', 'inputnode.t1_preproc'),
             ('t1_mask', 'inputnode.t1_mask'),
             ('t1_seg', 'inputnode.t1_seg'),
+            ('t1_tpms', 'inputnode.t1_tpms'),
             ('t1_2_mni_forward_transform', 'inputnode.t1_2_mni_forward_transform'),
             ('t1_2_mni', 'inputnode.t1_2_mni'),
             ('mni_mask', 'inputnode.mni_mask'),
+            ('mni_seg', 'inputnode.mni_seg'),
             ('mni_tpms', 'inputnode.mni_tpms'),
             ('surfaces', 'inputnode.surfaces'),
             ]),
@@ -398,39 +410,32 @@ def init_autorecon_resume_wf(omp_nthreads, name='autorecon_resume_wf'):
         name='outputnode')
 
     autorecon2_vol = pe.Node(
-        fs.ReconAll(
-            directive='autorecon2-volonly',
-            openmp=omp_nthreads),
+        fs.ReconAll(directive='autorecon2-volonly', openmp=omp_nthreads),
+        n_procs=omp_nthreads,
         name='autorecon2_vol')
-    autorecon2_vol.interface.num_threads = omp_nthreads
-
-    autorecon2_surfs = pe.MapNode(
-        fs.ReconAll(
-            directive='autorecon2-perhemi',
-            openmp=omp_nthreads),
-        iterfield='hemi',
-        name='autorecon2_surfs')
-    autorecon2_surfs.interface.num_threads = omp_nthreads
-    autorecon2_surfs.inputs.hemi = ['lh', 'rh']
 
     autorecon_surfs = pe.MapNode(
         fs.ReconAll(
             directive='autorecon-hemi',
-            flags=['-noparcstats', '-noparcstats2', '-noparcstats3',
-                   '-nohyporelabel', '-nobalabels'],
+            flags=['-noparcstats', '-nocortparc2', '-noparcstats2',
+                   '-nocortparc3', '-noparcstats3', '-nopctsurfcon',
+                   '-nohyporelabel', '-noaparc2aseg', '-noapas2aseg',
+                   '-nosegstats', '-nowmparc','-nobalabels'],
             openmp=omp_nthreads),
-        iterfield='hemi',
+        iterfield='hemi', n_procs=omp_nthreads,
         name='autorecon_surfs')
-    autorecon_surfs.interface.num_threads = omp_nthreads
     autorecon_surfs.inputs.hemi = ['lh', 'rh']
 
-    autorecon3 = pe.Node(
-        ReconAllRPT(
-            directive='autorecon3',
-            openmp=omp_nthreads,
-            generate_report=True),
+    autorecon3 = pe.MapNode(
+        fs.ReconAll(directive='autorecon3', openmp=omp_nthreads),
+        iterfield='hemi', n_procs=omp_nthreads,
         name='autorecon3')
-    autorecon3.interface.num_threads = omp_nthreads
+    autorecon3.inputs.hemi = ['lh', 'rh']
+
+    # Only generate the report once; should be nothing to do
+    recon_report = pe.Node(
+        ReconAllRPT(directive='autorecon3', generate_report=True),
+        name='recon_report')
 
     def _dedup(in_list):
         vals = set(in_list)
@@ -440,18 +445,18 @@ def init_autorecon_resume_wf(omp_nthreads, name='autorecon_resume_wf'):
         return vals.pop()
 
     workflow.connect([
-        (inputnode, autorecon_surfs, [('use_T2', 'use_T2')]),
+        (inputnode, autorecon3, [('use_T2', 'use_T2')]),
         (inputnode, autorecon2_vol, [('subjects_dir', 'subjects_dir'),
                                      ('subject_id', 'subject_id')]),
-        (autorecon2_vol, autorecon2_surfs, [('subjects_dir', 'subjects_dir'),
+        (autorecon2_vol, autorecon_surfs, [('subjects_dir', 'subjects_dir'),
                                             ('subject_id', 'subject_id')]),
-        (autorecon2_surfs, autorecon_surfs, [(('subjects_dir', _dedup), 'subjects_dir'),
-                                             (('subject_id', _dedup), 'subject_id')]),
         (autorecon_surfs, autorecon3, [(('subjects_dir', _dedup), 'subjects_dir'),
                                        (('subject_id', _dedup), 'subject_id')]),
-        (autorecon3, outputnode, [('subjects_dir', 'subjects_dir'),
-                                  ('subject_id', 'subject_id'),
-                                  ('out_report', 'out_report')]),
+        (autorecon3, outputnode, [(('subjects_dir', _dedup), 'subjects_dir'),
+                                  (('subject_id', _dedup), 'subject_id')]),
+        (autorecon3, recon_report, [(('subjects_dir', _dedup), 'subjects_dir'),
+                                    (('subject_id', _dedup), 'subject_id')]),
+        (recon_report, outputnode, [('out_report', 'out_report')]),
         ])
 
     return workflow
@@ -606,22 +611,28 @@ def init_anat_derivatives_wf(output_dir, output_spaces, template, freesurfer,
 
     inputnode = pe.Node(
         niu.IdentityInterface(
-            fields=['source_file', 't1_preproc', 't1_mask', 't1_seg',
+            fields=['source_file', 't1_preproc', 't1_mask', 't1_seg', 't1_tpms',
                     't1_2_mni_forward_transform', 't1_2_mni', 'mni_mask',
-                    'mni_tpms', 'surfaces']),
+                    'mni_seg', 'mni_tpms', 'surfaces']),
         name='inputnode')
 
     ds_t1_preproc = pe.Node(
         DerivativesDataSink(base_directory=output_dir, suffix='preproc'),
         name='ds_t1_preproc', run_without_submitting=True)
 
+    ds_t1_mask = pe.Node(
+        DerivativesDataSink(base_directory=output_dir, suffix='brainmask'),
+        name='ds_t1_mask', run_without_submitting=True)
+
     ds_t1_seg = pe.Node(
         DerivativesDataSink(base_directory=output_dir, suffix='dtissue'),
         name='ds_t1_seg', run_without_submitting=True)
 
-    ds_t1_mask = pe.Node(
-        DerivativesDataSink(base_directory=output_dir, suffix='brainmask'),
-        name='ds_t1_mask', run_without_submitting=True)
+    ds_t1_tpms = pe.Node(
+        DerivativesDataSink(base_directory=output_dir,
+                            suffix='class-{extra_value}_probtissue'),
+        name='ds_t1_tpms', run_without_submitting=True)
+    ds_t1_tpms.inputs.extra_values = ['CSF', 'GM', 'WM']
 
     suffix_fmt = 'space-{}_{}'.format
     ds_t1_mni = pe.Node(
@@ -633,6 +644,11 @@ def init_anat_derivatives_wf(output_dir, output_spaces, template, freesurfer,
         DerivativesDataSink(base_directory=output_dir,
                             suffix=suffix_fmt(template, 'brainmask')),
         name='ds_mni_mask', run_without_submitting=True)
+
+    ds_mni_seg = pe.Node(
+        DerivativesDataSink(base_directory=output_dir,
+                            suffix=suffix_fmt(template, 'dtissue')),
+        name='ds_mni_seg', run_without_submitting=True)
 
     ds_mni_tpms = pe.Node(
         DerivativesDataSink(base_directory=output_dir,
@@ -667,6 +683,8 @@ def init_anat_derivatives_wf(output_dir, output_spaces, template, freesurfer,
                                  ('t1_mask', 'in_file')]),
         (inputnode, ds_t1_seg, [('source_file', 'source_file'),
                                 ('t1_seg', 'in_file')]),
+        (inputnode, ds_t1_tpms, [('source_file', 'source_file'),
+                                 ('t1_tpms', 'in_file')]),
         ])
 
     if freesurfer:
@@ -684,6 +702,8 @@ def init_anat_derivatives_wf(output_dir, output_spaces, template, freesurfer,
                                     ('t1_2_mni', 'in_file')]),
             (inputnode, ds_mni_mask, [('source_file', 'source_file'),
                                       ('mni_mask', 'in_file')]),
+            (inputnode, ds_mni_seg, [('source_file', 'source_file'),
+                                      ('mni_seg', 'in_file')]),
             (inputnode, ds_mni_tpms, [('source_file', 'source_file'),
                                       ('mni_tpms', 'in_file')]),
             ])