+++ CHANGES.rst
@@ -1,7 +1,8 @@
 Next release
 ============
 
-
+* [ENH] Add --anat-only workflow option (#560)
+* [FIX] Output all tissue class/probability maps (#569)
 
 0.4.6 (14th of June 2017)
 =========================